Factoring out hosting defaults to separate class
